+++ hicn-light/src/hicn/base/test/test-pool.cc
@@ -56,7 +56,7 @@ TEST_F(PoolTest, PoolAllocation)
     /* Check that free indices and bitmaps are correctly initialize */
     off_t * fi = pool_get_free_indices(pool);
     EXPECT_EQ(vector_len(fi), pool_size);
-    EXPECT_EQ(fi[0], pool_size - 1);
+    EXPECT_EQ(fi[0], (long) (pool_size - 1));
     EXPECT_EQ(fi[pool_size - 1], 0);
 
     /* The allocated size of the underlying vector should be the next power of two */
@@ -89,12 +89,12 @@ TEST_F(PoolTest, PoolAllocation)
 
     rc = pool_get(pool, elt);
     EXPECT_GE(rc, 0);
-    EXPECT_EQ(vector_len(fi), 1);
+    EXPECT_EQ(vector_len(fi), 1UL);
     EXPECT_TRUE(bitmap_is_unset(fb, pool_size - 2));
 
     rc = pool_get(pool, elt);
     EXPECT_GE(rc, 0);
-    EXPECT_EQ(vector_len(fi), 0);
+    EXPECT_EQ(vector_len(fi), 0UL);
     EXPECT_TRUE(bitmap_is_unset(fb, pool_size - 1));
 
     /*
